Search Page basic front-end functions implemented, needs debugging
If i return GenerateResults as a component it will basically be called every single time i type something because of the onChange setName function. This back and forth somehow calls the backend exactly 3 times. It is able to retrieve data in 2 of these calls but fails once. The failed call is what gets stored in the array. No error is reported to the console though for whatever reason despite the "console.error that I wrote".